packages/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["MichaelDickinson20I <michael.dickinson3949@gmail.com>"]
readme = "README.md"
packages = [
  { include = "*", from = "src" }
]


[tool.poetry.dependencies]
python = "^3.12"
python-docx = "^1.1.2"
firebase-admin = "^6.8.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"
pydantic = "^2.11.4"
md2docx-python = "^1.0.0"
google-genai = "^1.14.0"
python-dotenv = "^1.1.0"
firebase-functions = "^0.4.2"
cloudconvert = "^2.1.0"
anthropic = "^0.52.1"
numpy = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

